added known fingerprints to O2-connector.
This is a fix for the man-in-the-middle-possibility introduced for the hotfix
of issue #272.
It needs commit 1be6bdba374410a9f67c from my fork of websms-api to work.

Please check the fingerprints of login.o2online.de and email.o2online.de
yourself before pulling, don't trust me on this ;-)